backend/pyproject.toml
[project]
name = "neil-solar-api"
version = "0.1.0"
description = "All Solar AMC SaaS API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "alembic>=1.13.2",
  "python-multipart>=0.0.9",
  "google-auth>=2.30.0",
  "requests>=2.32.0",
]

[tool.alembic]
script_location = "alembic"

[tool.ruff]
line-length = 100